Move orion.algo.parallel_strategy to orion.algo.base
