+++ src/orion/algo/base/parallel_strategy.py
