+++ src/orion/algo/tpe.py
@@ -12,7 +12,7 @@ from scipy.stats import norm
 from scipy.stats.distributions import rv_frozen
 
 from orion.algo.base import BaseAlgorithm
-from orion.algo.parallel_strategy import ParallelStrategy, strategy_factory
+from orion.algo.base.parallel_strategy import ParallelStrategy, strategy_factory
 from orion.algo.space import Integer, Real, Space
 from orion.core.utils import format_trials
 from orion.core.worker.trial import Trial
